Bugfixes with towers and ore tiles
item sprites scrolling across conveyors were way too small
Spent ages fixing a bug that caused conveyors to break when placed on or next to an ore tile. seriously, I am not proud of how long it took to fix this.
simplified some code in conveyor code to aid readability of code
added a Tile observer to aid in debugging: you may no click a tile in the scene and view all a tile's values in the inspector. Aside from debugging, this is  fairly useless code, so there is a bool in MapGeneration.Cs to disable this.